+++ controllers/serviceController.js
@@ -1,4 +1,7 @@
 const Service = require('./../models/serviceModel');
+const CallHistory = require('./../models/callHistoryModel');
+const CollaboratorHistory = require('./../models/collaboratorHistoryModel');
+const ServiceIntercom = require('./../models/serviceIntercomModel');
 
 exports.getAllServices = async (req, res) => {
     try {
@@ -61,10 +64,21 @@ exports.updateService = async (req, res) => {
 exports.deleteService = async (req, res) => {
 
     try {
-        await Service.findByIdAndDelete(req.params.id)
-        res.status(204).json({
-            status: 'succes'
-        })
+        const callHistory = await CallHistory.findOne({ service: req.params.id });
+        const collaboratorHistory = await CollaboratorHistory.findOne({ service: req.params.id });
+        const serviceIntercom = await ServiceIntercom.findOne({ service: req.params.id });
+
+        if (callHistory || collaboratorHistory || serviceIntercom) {
+            res.status(200).json({
+                status: 'stopped',
+                message: `This collaborator is used in another entry. It coudldn't be deleted.`
+            });
+        } else {
+            await Service.findByIdAndDelete(req.params.id)
+            res.status(204).json({
+                status: 'succes'
+            })
+        }
     }
     catch (err) {
         res.status(404).json({
